refactor generating token
