+++ src/models/project.ts
@@ -2,6 +2,7 @@ import { Collection, ObjectId } from 'mongodb';
 import AbstractModel from './abstractModel';
 import { NotificationsChannelsDBScheme } from '../types/notification-channels';
 import { ProjectDBScheme } from 'hawk.types';
+import uuid from "uuid";
 
 /**
  * This structure represents a single rule of notifications settings
@@ -189,6 +190,24 @@ export default class ProjectModel extends AbstractModel<ProjectDBScheme> impleme
   constructor(projectData: ProjectDBScheme) {
     super(projectData);
     this.collection = this.dbConnection.collection<ProjectDBScheme>('projects');
+  }
+
+  /**
+   * Generates new integration token with integration id field
+   *
+   * @param integrationId - integration id for using in collector URL
+   */
+  public static generateIntegrationToken(integrationId: string): string {
+    const randomHash = uuid.v4();
+
+    const decodedIntegrationToken = {
+      integrationId,
+      randomHash,
+    };
+
+    return Buffer
+      .from(JSON.stringify(decodedIntegrationToken))
+      .toString('base64');
   }
 
   /**
